Add files via upload
Updated code to support the RP2040, ADG6412, and FS58R3MW

+++ src/rx5808-pro-diversity/timer.cpp
@@ -2,7 +2,9 @@
 #include "timer.h"
 
 
-Timer::Timer(uint16_t delay) {
+// PICO PORT: Upgraded delay to uint32_t.
+// This prevents overflow for timers longer than 65 seconds (uint16 limit).
+Timer::Timer(uint32_t delay) {
     this->delay = delay;
     this->nextTick = millis() + this->delay;
     this->ticked = false;
