+++ src/rx5808-pro-diversity/timer.h
@@ -8,11 +8,11 @@
 class Timer {
     private:
         uint32_t nextTick;
-        uint16_t delay;
+        uint32_t delay; // PICO PORT: Changed from uint16_t to uint32_t
         bool ticked;
 
     public:
-        Timer(uint16_t delay);
+        Timer(uint32_t delay);
         const bool hasTicked();
         void reset();
 };
